Move menu-item directory into menu
